pyproject.toml
[tool.poetry]
name = "my-library"
version = "0.1.0"
description = ""
authors = ["lisvv <81390205+lisvv@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
drf-spectacular = "^0.26.2"
django-celery-beat = "^2.5.0"
python-dateutil = "^2.8.2"
redis = "^4.5.5"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
